DAQ-527 Make Spring AOP classes visible to Jython bundle
When Jython encounters a bean that uses Spring AOP it actually gets a
proxy object that refers to the real one via several classes and so has
to make python representations of each object type. This is done by the
Jython bundle (not the interpreter we create) and so it uses the Jython
bundle's classLoader which can't see Spring AOP classes.

Add a dependency on uk.ac.diamond.org.spring.framework (which contains
the AOP jar) to solve this and any other potential future references to
Spring classes.