DAQ-612 Define Fragment to add Spring classes to Jython bundle classpath
Jython (the jar within the Jython bundle) needs to be able to load the
Spring AOP classes in order to be able to load beans based on classes
that make use of these facilities. Previously a hard dependency on
Spring was added but this is undesirable, so instead use the OSGi
recommended method of providing a fragment hosted by Jython that depends
on the Spring classes. 

This allows the existing Jython plugin to be used in isolation as before
but allows the GDA server to give access to the Spring classes by
including the fragment in one of its features.