DAQ-1032 Switch labelling sense of split packages
The idea is to change the places where split packages need to be
identified. Previous to this change everywhere a split package was
exported it needed to be labelled as split so that the OSGi wouldn't
wire it to the Jython bundle. After this change it will be
preferentially wired to exports specifying they are merged.

This means only the manifest of the split package resolver need to know
about this issue and other bundle manifests exporting split packages can
be left untouched.

Change-Id: Iec3cc6eee798913c93d3b87bf06ddaff517077e8