DAQ-3351 Workaround for restricted permissions of py class files
Jython now creates class files with u+rw permissions only (600). This
means that they cannot be read by other users.

This is a problem for shared workspace deployments as Jython is invoked
by Pydev in the IDE creating these files in
uk.ac.diamond.jython/jython2.7/Lib

When the plugin is build later on the contents of
uk.ac.diamond.jython/jython2.7/Lib needs to be exported but if a
different user is running the build it will fail as the *$py.class
files are not readable.

This change excludes *$py.class from the binary build.

Ideally we would find a way to stop these files being written in the
first place.

Change-Id: Idbe01570cba96e444671886ee38e13cf8a8c7f59